requirements.txt
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.1
six==1.14.0
torch==1.4.0+cpu
torchvision==0.5.0+cpu
